#roxmltree committed to string based XmlStream and removed byte based XmlStream
