+++ packages/roxmltree/src/parse-string.ts
@@ -1,9 +1,9 @@
-import { parseXmlStream, TextXmlStream, type TTokenCallback } from './tokenizer';
+import { parseXmlStream, XmlStream, type TTokenCallback } from './tokenizer';
 
 export function parseString(
 	xmlString: string,
 	allowDtd: boolean,
 	tokenCallback: TTokenCallback
 ): void {
-	parseXmlStream(new TextXmlStream(xmlString), allowDtd, tokenCallback);
+	parseXmlStream(new XmlStream(xmlString), allowDtd, tokenCallback);
 }
