--- packages/roxmltree/src/tokenizer/XMLStream/index.ts
@@ -1,3 +1,0 @@
-export * from './ByteXmlStream';
-export * from './TextXmlStream';
-export * from './types';
