+++ packages/roxmltree/src/tokenizer/tokenizer.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest';
 import { parseXmlStream } from './tokenizer';
 import { type TXMLToken } from './types';
 import { XmlError } from './XmlError';
-import { ByteXmlStream, TextXmlStream } from './XmlStream';
+import { XmlStream } from './XmlStream';
 
 describe('tokenizer tests', () => {
 	describe('CDATA', () => {
@@ -2243,9 +2243,8 @@ interface TAttrToken {
 function collectTokens(text: string): TToken[] {
 	const tokens: TToken[] = [];
 	try {
-		const textXmlStream = new TextXmlStream(text);
-		const byteXmlStream = new ByteXmlStream(new TextEncoder().encode(text));
-		parseXmlStream(textXmlStream, true, (token) => {
+		const xmlStream = new XmlStream(text);
+		parseXmlStream(xmlStream, true, (token) => {
 			switch (token.type) {
 				case 'EntityDeclaration':
 					tokens.push({
